feat(notifications): converted component to typescript
